d13: with some prodding, get part a
the issue was i was too generous, i misread/misunderstood the
question to be not exit on a 'correct' comparison. My first attempt
required that every component of the packet be correct in order to
be correct, not just the first comparison that was correct signifying
for the whole packet.